@@ -17,22 +17,22 @@
 		"createAdmin": "npx tsx src/scripts/createAdmin.ts",
 		"createOrganization": "npx tsx src/scripts/createOrganization.ts",
 		"createKeypair": "node src/scripts/createKeypair",
-		"createS3BucketLocalStack": "npx tsx src/scripts/createS3BucketLocalStack.ts"
+		"createS3BucketLocalStack": "npx tsx src/scripts/createS3BucketLocalStack.ts",
+		"machine-translate": "inlang machine translate --project project.inlang"
 	},
 	"devDependencies": {
 		"@digitalbazaar/ed25519-signature-2020": "^5.1.0",
 		"@digitalbazaar/ed25519-verification-key-2020": "^4.1.0",
 		"@digitalbazaar/vc": "^6.0.0",
-		"@inlang/cli": "^2.0.0",
-		"@inlang/paraglide-js": "^1.3.3",
-		"@inlang/paraglide-sveltekit": "^0.12.1",
+		"@inlang/cli": "^3.0.10",
+		"@inlang/paraglide-js": "2.0.10",
 		"@inlang/plugin-message-format": "^2.0.0",
 		"@playwright/test": "^1.40.0",
 		"@popperjs/core": "^2.11.8",
-		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/adapter-auto": "^4.0.0",
 		"@sveltejs/kit": "^2.9.0",
-		"@sveltejs/package": "^2.2.3",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/package": "^2.3.7",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
 		"@types/http-link-header": "^1.0.5",
 		"@types/inquirer": "^9.0.3",
 		"@types/mime-types": "^2.1.1",
@@ -64,20 +64,20 @@
 		"postcss-load-config": "^4.0.1",
 		"prettier": "^3.1.0",
 		"prettier-plugin-prisma": "^5.0.0",
-		"prettier-plugin-svelte": "^3.1.2",
-		"prisma": "^5.7.1",
+		"prettier-plugin-svelte": "^3.2.6",
+		"prisma": "^6.5.0",
 		"sanitize-html": "^2.9.0",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.0.3",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"svelte-droplet": "^1.0.0",
 		"svelte-icons-pack": "^2.1.0",
-		"svelte-meta-tags": "^3.1.0",
-		"svelte-preprocess": "^5.0.1",
+		"svelte-meta-tags": "^4.2.0",
+		"svelte-preprocess": "^6.0.0",
 		"tailwindcss": "^3.3.6",
-		"typescript": "^5.3.2",
+		"typescript": "^5.5.0",
 		"uuid": "^9.0.0",
-		"vite": "^5.0.0",
-		"vite-plugin-static-copy": "^0.17.1",
+		"vite": "^6.0.0",
+		"vite-plugin-static-copy": "^2.3.0",
 		"vitest": "^3.0.5",
 		"vitest-mock-extended": "^2.0.2",
 		"yup": "^1.0.0"
@@ -86,18 +86,22 @@
 	"dependencies": {
 		"@aws-sdk/client-s3": "^3.299.0",
 		"@aws-sdk/s3-request-presigner": "^3.326.0",
-		"@neondatabase/serverless": "^0.6.1",
-		"@prisma/adapter-neon": "^5.7.1",
-		"@prisma/client": "^5.7.1",
+		"@neondatabase/serverless": "^1.0.0",
+		"@prisma/adapter-neon": "^6.5.0",
+		"@prisma/client": "^6.5.0",
 		"@vitejs/plugin-basic-ssl": "^1.0.1",
-		"carta-md": "^3.4.0",
+		"carta-md": "^4.8.1",
 		"credential-handler-polyfill": "^3.2.0",
 		"dayjs": "^1.11.9",
 		"jsonld-signatures": "^11.2.0",
+		"marked": "^15.0.7",
 		"mime-types": "^2.1.35",
 		"nodemailer-mailgun-transport": "^2.1.5",
 		"qrcode": "^1.5.3",
 		"ws": "^8.16.0"
+	},
+	"engines": {
+		"node": "22.x"
 	},
 	"prisma": {
 		"schema": "src/prisma/schema.prisma"
