@@ -1,5 +1,6 @@
 <script lang="ts">
   import * as m from "$lib/i18n/messages";
+  import type { Snippet } from "svelte";
 
   interface Item {
     text: string;
@@ -9,7 +10,12 @@
     class?: string;
   }
 
-  export let items: Array<Item>;
+  interface Props {
+    items: Array<Item>;
+    divider?: Snippet;
+  }
+
+  let { items, divider }: Props = $props();
 </script>
 
 <nav class="flex" aria-label={m.breadcrumb()}>
@@ -18,28 +24,26 @@
       {#if i !== 0}
         <li class="divider dark:text-gray-400">
           <!-- The slot used for divider -->
-          <slot name="divider">/</slot>
+          {#if divider}{@render divider()}{:else}/{/if}
         </li>
       {/if}
       <li class="inline-flex items-center">
-        <slot {item}>
-          {#if item.href}
-            <a
-              href={item.href}
-              class="breadcrumb-item inline-flex items-center text-sm font-medium text-gray-700 hover:text-gray-900 dark:text-gray-300 dark:hover:text-white {item.class}"
-              {...item.props}
-            >
-              {item.text}
-            </a>
-          {:else}
-            <span
-              class="breadcrumb-item disabled text-sm font-medium text-gray-500 dark:text-gray-400 {item.class}"
-              {...item.props}
-            >
-              {item.text}
-            </span>
-          {/if}
-        </slot>
+        {#if item.href}
+          <a
+            href={item.href}
+            class="breadcrumb-item inline-flex items-center text-sm font-medium text-gray-700 hover:text-gray-900 dark:text-gray-300 dark:hover:text-white {item.class}"
+            {...item.props}
+          >
+            {item.text}
+          </a>
+        {:else}
+          <span
+            class="breadcrumb-item disabled text-sm font-medium text-gray-500 dark:text-gray-400 {item.class}"
+            {...item.props}
+          >
+            {item.text}
+          </span>
+        {/if}
       </li>
     {/each}
   </ol>
