@@ -1,12 +1,23 @@
 <script lang="ts">
-	let klass = 'bg-slate-100 text-slate-700';
-	export { klass as class };
-	export let open = true;
+	
+	interface Props {
+		class?: string;
+		open?: boolean;
+		trigger?: import('svelte').Snippet;
+		children?: import('svelte').Snippet;
+	}
+
+	let {
+		class: klass = 'bg-slate-100 text-slate-700',
+		open = $bindable(true),
+		trigger,
+		children
+	}: Props = $props();
 </script>
 
 <span class={`whitespace-nowrap rounded-lg py-0.5 px-2 text-xs font-semibold leading-6 ${klass}`}>
-	<button on:click={() => (open = !open)}> <slot name="trigger" /></button>
+	<button onclick={() => (open = !open)}> {@render trigger?.()}</button>
 	{#if open}
-		<slot />
+		{@render children?.()}
 	{/if}
 </span>
