@@ -1,13 +1,23 @@
 <script lang="ts">
-	export let text = '';
-	let inputId: string;
-	export { inputId as for };
-	export let disabled = false;
+	
+	interface Props {
+		text?: string;
+		for: string;
+		disabled?: boolean;
+		children?: import('svelte').Snippet;
+	}
+
+	let {
+		text = '',
+		for: inputId,
+		disabled = false,
+		children
+	}: Props = $props();
 
 	const tColor = 'text-gray-900 dark:text-gray-300'; // Default text color
 	const dtColor = 'text-gray-400 dark:text-gray-600'; // Disabled text color
 </script>
 
 <label for={inputId} class={`block mb-2 text-sm font-medium ${disabled ? dtColor : tColor}`}>
-	<slot>{text}</slot>
+	{#if children}{@render children()}{:else}{text}{/if}
 </label>
