@@ -6,7 +6,11 @@
 	import { calculatePageAndSize } from '$lib/utils/pagination';
 	import type { PageData } from './$types';
 
-	export let data: PageData;
+	interface Props {
+		data: PageData;
+	}
+
+	let { data }: Props = $props();
 	let members = data.members;
 </script>
 
