@@ -35,7 +35,7 @@ declare namespace App {
 		org: Organization;
 		token: string;
 		session: SessionData?;
-		locale: import('$lib/i18n/runtime').AvailableLanguageTag;
+		locale: string;
 	}
 	interface PageData {
 		org: Organization;
