@@ -1,4 +1,4 @@
-import { derived, writable } from 'svelte/store';
+import { writable } from 'svelte/store';
 import { LoadingStatus } from './common';
 
 export const session = writable<App.SessionData | undefined>();
